requirements.txt
# Core framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pgvector==0.2.4

# AI/ML
openai==1.10.0
tiktoken==0.5.2
langchain==0.1.4
langchain-openai==0.0.5

# Data processing
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2

# HTTP client (for future Zendesk API integration)
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
